requirements.txt
anthropic
python-dotenv
pypdf
streamlit
python-docx
pdf2docx